Add support for running on windows
This commit switches to using subprocess run without depending on shell
semantics and weird quoting which didn't work on windows. Also it times
the run inside python as on windows the time executable is used to set
the system time, not to time something so it didn't work. Using a ns
accurate perf counter we could also increase the precision of the timing
as needed.