Show correct test count, restructure available suites
We were showing the total number of files in each suite directory,
regardless of the amount of tests we were actually going to execute.

Register a list of available suites with their respective suffix and
other configuration so we can generalize the code a bit, allowing us to
produce the correct total number of tests.